Move `passives.ts`
