+++ src/field/pokemon.ts
@@ -51,7 +51,7 @@ import type { Ability } from "#app/data/abilities/ability";
 import { applyAbAttrs, getAbApplyFunc } from "#app/data/abilities/apply-ab-attrs";
 import { NoCritTag } from "#app/data/arena-tag";
 import { speciesEggMoves } from "#app/data/egg-moves";
-import { starterPassiveAbilities } from "#app/data/balance/passives";
+import { starterPassiveAbilities } from "#app/data/passives";
 import { pokemonEvolutions } from "#app/data/pokemon-evolutions/init-pokemon-evolutions";
 import { EVOLVE_MOVE, RELEARN_MOVE, type LevelMoves } from "#app/data/balance/pokemon-level-moves";
 import {
